requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.5
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
python-dotenv==1.0.0
pyttsx3==2.90
requests==2.29.0
tomli==2.0.1
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
yarl==1.9.2
